# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for managing todos with JWT authentication",
        "title": "GoTodo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/anchors/{id}/follow": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Follow or unfollow an anchor to track updates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anchor-follows"
                ],
                "summary": "Follow or unfollow an anchor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Anchor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Follow action",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/anchor_follows.FollowAnchorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/anchor_follows.FollowAnchorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/anchors/{id}/follow/notifications": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enable or disable notifications when anchor is updated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anchor-follows"
                ],
                "summary": "Toggle update notifications for followed anchor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Anchor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notification preference",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/anchor_follows.ToggleNotificationsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/anchors/{id}/follow/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check if current user is following an anchor and get update status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anchor-follows"
                ],
                "summary": "Get follow status for an anchor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Anchor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/anchor_follows.FollowStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/users/me/following-anchors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all anchors the current user is following with update status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anchor-follows"
                ],
                "summary": "List anchors user is following",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter only anchors with updates",
                        "name": "hasUpdates",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort: recent, updated, alphabetical",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/anchor_follows.ListFollowingAnchorsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "anchor_follows.AuthorInfo": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "profilePictureUrl": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "anchor_follows.FollowAnchorRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "follow",
                        "unfollow"
                    ]
                },
                "notifyOnUpdate": {
                    "type": "boolean"
                }
            }
        },
        "anchor_follows.FollowAnchorResponse": {
            "type": "object",
            "properties": {
                "followerCount": {
                    "type": "integer"
                },
                "isFollowing": {
                    "type": "boolean"
                },
                "notifyOnUpdate": {
                    "type": "boolean"
                }
            }
        },
        "anchor_follows.FollowStatusResponse": {
            "type": "object",
            "properties": {
                "currentVersion": {
                    "type": "integer"
                },
                "followedAt": {
                    "type": "string"
                },
                "hasUpdates": {
                    "type": "boolean"
                },
                "isFollowing": {
                    "type": "boolean"
                },
                "lastSeenVersion": {
                    "type": "integer"
                },
                "notifyOnUpdate": {
                    "type": "boolean"
                },
                "updatesSinceLastSeen": {
                    "type": "integer"
                }
            }
        },
        "anchor_follows.FollowingAnchorItem": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/anchor_follows.AuthorInfo"
                },
                "currentVersion": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "followedAt": {
                    "type": "string"
                },
                "followerCount": {
                    "type": "integer"
                },
                "hasUpdates": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "itemCount": {
                    "type": "integer"
                },
                "lastItemAddedAt": {
                    "type": "string"
                },
                "lastSeenVersion": {
                    "type": "integer"
                },
                "likeCount": {
                    "type": "integer"
                },
                "notifyOnUpdate": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatesSinceLastSeen": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "anchor_follows.ListFollowingAnchorsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/anchor_follows.FollowingAnchorItem"
                    }
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "sort": {
                            "type": "string"
                        },
                        "totalWithUpdates": {
                            "type": "integer"
                        }
                    }
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "hasMore": {
                            "type": "boolean"
                        },
                        "limit": {
                            "type": "integer"
                        },
                        "page": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        },
                        "totalPages": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "anchor_follows.ToggleNotificationsRequest": {
            "type": "object",
            "properties": {
                "notifyOnUpdate": {
                    "type": "boolean"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH_INVALID_TOKEN"
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}